+++ tools/config/src/config_cli.rs
@@ -1,7 +1,7 @@
 use crate::{
     config_wizard,
     get_genesis_artifacts::{download_genesis, get_genesis_waypoint},
-    interactive,
+    interactive::options::{fix_config, FixOptions},
     make_yaml_public_fullnode::init_fullnode_yaml,
     validator_config::{validator_dialogue, vfn_dialogue},
 };
@@ -95,7 +95,7 @@ impl ConfigCli {
                 remove_profile,
                 fullnode_url: force_url,
             }) => {
-                interactive::fix_config(interactive::FixOptions {
+                fix_config(FixOptions {
                     reset_address: *reset_address,
                     remove_profile: remove_profile.clone(),
                     fullnode_url: force_url.clone(),
