+++ tools/config/src/interactive/mod.rs
@@ -2,11 +2,8 @@
 
 pub mod account_selection;
 pub mod defaults;
-pub mod interactive;
+pub mod fix;
 pub mod networks;
 pub mod options;
 pub mod profiles;
 pub mod utils;
-
-// Re-export the main public interface
-pub use options::{fix_config, FixOptions};
